requirements.txt
fastapi
uvicorn[standard]
google-cloud-vision
python-multipart
requests
